consumer/requirements.txt
pika
mysql-connector-python
opentelemetry-distro
opentelemetry-exporter-otlp
